requirements.txt
flask
flask-login
flask-sqlalchemy
werkzeug
opencv-python